quickfixhub-backend/requirements.txt
pymongo
dnspython
Flask
Flask-Login
Flask-Cors
Flask-Bcrypt

boto3
botocore

python-dotenv